+++ llm/config.py
@@ -10,7 +10,9 @@ ENABLED_PROVIDER = os.getenv("ENABLED_PROVIDER", "openai").split(",")
 PROXY_SERVER = os.getenv("PROXY_SERVER", None)
 
 NO_GUI = os.getenv("NO_GUI", "false").lower() == "true"
-HEADLESS = os.getenv("HEADLESS", "true").lower() == "true"
+SCREEN_WIDTH = int(os.getenv("SCREEN_WIDTH", "1080"))
+SCREEN_HEIGHT = int(os.getenv("SCREEN_HEIGHT", "1920"))
+HEADLESS = os.getenv("HEADLESS", "false").lower() == "true"
 USER_AGENT = os.getenv("USER_AGENT", None)
 SAVE_LOGIN_STATE = os.getenv("SAVE_LOGIN_STATE", True)
 
@@ -20,8 +22,6 @@ os.makedirs(BROWSER_DATA, exist_ok=True)
 
 env = os.getenv("env", "prod")
 
-SCREEN_WIDTH = 1080
-SCREEN_HEIGHT = 1920
 
 OPENAI_LOGIN_TYPE = os.getenv("OPENAI_LOGIN_TYPE", "nologin")
 OPENAI_LOGIN_EMAIL = os.getenv("OPENAI_LOGIN_EMAIL", "")